164892850-feature(auth) refactor
-refactor code

[delivers  #164892850]
