+++ src/routers/index.js
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 import passport from 'passport';
 import { AddArticles, UpdateArticle, DeleteArticle } from '../controllers/article';
-import user from './user';
+
 
 const router = Router();
 
@@ -20,7 +20,7 @@ router.get('/', (req, res) => res.status(200).json({
  * @param {function} UpdateArticle - Express path
  * @returns Response Object
  */
-router.use('/', user);
+
 router
   .route('/articles/:id?')
   .post(AddArticles)
